Commit before removing the old training system

+++ src/network/fmt.rs
@@ -20,7 +20,7 @@ Network:
     - mode: {:?}
         ",
             self.inputs_count,
-            self.nodes.len() as i32 - self.inputs_count - self.leaves_count,
+            self.nodes.len() - self.inputs_count - self.leaves_count,
             self.leaves_count,
             self.get_connection_count(),
             self.mode
